ansible/example/djangoimages/requirements.txt
#conda create -n django python=3.6.8
datetime
django-taggit

